moving the form and the status bar into their own components to simplify the post-a-job.js page
